refactor: move advice injector parsers into own module
